string_nconcat 002

+++ 0x0C-more_malloc_free/1-string_nconcat.c
@@ -10,7 +10,7 @@
  */
 char *string_nconcat(char *s1, char *s2, unsigned int n)
 {
-	unsigned int i = 0, len1 = 0, len2 = 0, size = 0;
+	int i = 0, len1 = 0, len2 = 0, size = 0;
 	char *ptr;
 
 	for (; (s1 && s1[len1]) || (s2 && s2[len2]); )
@@ -20,7 +20,7 @@ char *string_nconcat(char *s1, char *s2, unsigned int n)
 		if (s2 && s2[len2])
 			len2++;
 	}
-	size = len1 + (n >= len2 ? len2 : n) + 1;
+	size = len1 + ((int)n >= len2 ? len2 : (int)n) + 1;
 	ptr = malloc(sizeof(*ptr) * size);
 
 	if (!ptr)
